6_User_Auth_AI_ChatBot/requirements.txt
streamlit 
pymongo
langchain 
langchain-openai 
langgraph 
langgraph-checkpoint
langgraph-checkpoint-redis
dotenv
python-dotenv
